agents/python/requirements.txt
langchain==1.1.0
langchain-openai
httpx==0.28.1
httpcore==1.0.6
h11==0.14.0
sniffio==1.3.1
certifi==2024.7.4
python-dotenv==1.0.1
PyJWT[crypto]==2.10.1
idna==3.10
cryptography==42.0.8
cffi==1.17.1
python-dateutil==2.9.0.post0
jupyter
ipykernel
